moved enum to the common enums folder
